Changed the value of i to hold the size of unsigned long int

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -11,7 +11,7 @@ void print_binary(unsigned long int n)
 	int i, bit, c;
 
 	c = 0;
-	for (i = sizeof(int) * 8 - 1; i >= 0; i--)
+	for (i = sizeof(unsigned long int) * 8 - 1; i >= 0; i--)
 	{
 		bit = (n >> i) & 1;
 		if (bit == 1)
